+++ CHANGES.rst
@@ -20,6 +20,7 @@ Changelog
 - Use declareProtected(ModifyPortalContent) for methods on MeetingItem 'setCategory', 'setClassifier', 'setProposingGroup' and 'setProposingGroupWithGroupInCharge'
 - Fixed bug when an item is sent to another MeetingConfig and fails to be presented in a meeting because none is available, it crashed to render the portal_message
   if the destination MeetingConfig title contained special characters
+- Changed text of collective.messagesviewlet 'Holidays warning' message to use a less panicking content
 
 4.1.8 (2019-10-14)
 ------------------